♻️ refactor: swagger 문서 분리
