+++ server/src/rss/api-docs/createRss.api-docs.ts
@@ -1,0 +1,40 @@
+import { applyDecorators } from '@nestjs/common';
+import {
+  ApiBadRequestResponse,
+  ApiConflictResponse,
+  ApiCreatedResponse,
+  ApiOperation,
+} from '@nestjs/swagger';
+
+export function ApiCreateRss() {
+  return applyDecorators(
+    ApiOperation({
+      summary: 'RSS 등록 API',
+    }),
+    ApiCreatedResponse({
+      description: 'Created',
+      schema: {
+        properties: {
+          message: {
+            type: 'string',
+          },
+        },
+      },
+      example: {
+        message: '신청이 완료되었습니다.',
+      },
+    }),
+    ApiBadRequestResponse({
+      description: 'Bad Request',
+      example: {
+        message: '오류 메세지',
+      },
+    }),
+    ApiConflictResponse({
+      description: 'Conflict',
+      example: {
+        message: '이미 등록된 RSS URL입니다.',
+      },
+    }),
+  );
+}
